Address more feedback from code review
