+++ lib/utils/registry/registry_windows.go
@@ -32,34 +32,19 @@ import (
 // GetOrCreateRegistryKey loads or creates a registry key handle and passes it back
 func GetOrCreateRegistryKey(name string) (registry.Key, error) {
 	reg, err := registry.OpenKey(registry.CURRENT_USER, name, registry.QUERY_VALUE|registry.CREATE_SUB_KEY|registry.SET_VALUE)
-	if errors.Is(err, os.ErrNotExist) {
+	switch {
+	case errors.Is(err, os.ErrNotExist):
 		log.Debugf("Registry key %v doesn't exist, trying to create it", name)
 		reg, _, err = registry.CreateKey(registry.CURRENT_USER, name, registry.QUERY_VALUE|registry.CREATE_SUB_KEY|registry.SET_VALUE)
 		if err != nil {
 			log.Debugf("Can't create registry key %v: %v", name, err)
 			return reg, err
 		}
-	} else {
+	case err != nil:
 		log.Errorf("registry.OpenKey returned error: %v", err)
 		return reg, err
-	}
-	return reg, nil
-}
-
-// GetKey loads or creates a registry key handle and passes it back
-func GetKey(name string) (registry.Key, error) {
-	// now check for and create the individual session key
-	reg, err := registry.OpenKey(registry.CURRENT_USER, name, registry.QUERY_VALUE|registry.CREATE_SUB_KEY|registry.SET_VALUE)
-	if errors.Is(err, os.ErrNotExist) {
-		log.Debugf("Registry key %v doesn't exist, trying to create it", name)
-		reg, _, err = registry.CreateKey(registry.CURRENT_USER, name, registry.QUERY_VALUE|registry.CREATE_SUB_KEY|registry.SET_VALUE)
-		if err != nil {
-			log.Debugf("Can't create registry key %v: %v", name, err)
-			return reg, err
-		}
-	} else {
-		log.Errorf("registry.OpenKey returned error: %v", err)
-		return reg, err
+	default:
+		return reg, nil
 	}
 	return reg, nil
 }
